+++ Package.swift
@@ -18,7 +18,7 @@ let package = Package(
         )
     ],
     dependencies:[
-		.package(url:"https://github.com/tannerdsilva/CLMDB.git", "0.9.26"..<"0.9.31"),
+		.package(url:"https://github.com/tannerdsilva/CLMDB.git", branch:"master"),
 		.package(url:"https://github.com/tannerdsilva/rawdog.git", "22.0.0"..<"23.0.0"),
 		.package(url:"https://github.com/apple/swift-system.git", "1.0.0"..<"2.0.0"),
 		.package(url:"https://github.com/apple/swift-syntax.git", "603.0.0"..<"604.0.0")
@@ -36,6 +36,8 @@ let package = Package(
 				"CLMDB",
 				.product(name:"SystemPackage", package:"swift-system"),
 				.product(name:"RAW", package:"rawdog"),
+				.product(name:"RAW_chachapoly", package:"rawdog"),
+				.product(name:"RAW_blake2", package:"rawdog"),
 				"QuickLMDBMacros",
 				"QuickLMDBFunctionalInterop",
 			],